+++ src/VootTokenHook.php
@@ -9,6 +9,9 @@
 
 namespace SURFnet\VPN\Portal;
 
+use fkooman\OAuth\Client\OAuthClient;
+use fkooman\OAuth\Client\Provider;
+use fkooman\SeCookie\SessionInterface;
 use SURFnet\VPN\Common\Http\BeforeHookInterface;
 use SURFnet\VPN\Common\Http\Exception\HttpException;
 use SURFnet\VPN\Common\Http\RedirectResponse;
@@ -21,12 +24,35 @@ use SURFnet\VPN\Common\HttpClient\ServerClient;
  */
 class VootTokenHook implements BeforeHookInterface
 {
+    /** @var \fkooman\SeCookie\SessionInterface */
+    private $session;
+
     /** @var \SURFnet\VPN\Common\HttpClient\ServerClient */
     private $serverClient;
 
-    public function __construct(ServerClient $serverClient)
+    /** @var \fkooman\OAuth\Client\OAuthClient */
+    private $client;
+
+    /** @var \fkooman\OAuth\Client\Provider */
+    private $provider;
+
+    /** @var string */
+    private $vootUri;
+
+    /**
+     * @param \fkooman\SeCookie\SessionInterface          $session
+     * @param \SURFnet\VPN\Common\HttpClient\ServerClient $serverClient
+     * @param \fkooman\OAuth\Client\OAuthClient           $client
+     * @param \fkooman\OAuth\Client\Provider              $provider
+     * @param string                                      $vootUri
+     */
+    public function __construct(SessionInterface $session, ServerClient $serverClient, OAuthClient $client, Provider $provider, $vootUri)
     {
+        $this->session = $session;
         $this->serverClient = $serverClient;
+        $this->client = $client;
+        $this->provider = $provider;
+        $this->vootUri = $vootUri;
     }
 
     /**
@@ -47,6 +73,18 @@ class VootTokenHook implements BeforeHookInterface
         }
         $userInfo = $hookData['auth'];
 
+        // check if we have the group info in the cache
+        if ($this->session->has('_cached_voot_groups')) {
+            // does it match the current userId?
+            if ($userInfo->id() === $this->session->get('_cached_voot_groups_user_id')) {
+                $vootGroups = $this->session->get('_cached_voot_groups');
+                $hookData['auth']->addEntitlements($vootGroups);
+
+                return true;
+            }
+            // does not match expected user... fetch again!
+        }
+
         // do not get involved in POST requests, only in GETs
         if ('GET' !== $request->getRequestMethod()) {
             return false;
@@ -61,7 +99,15 @@ class VootTokenHook implements BeforeHookInterface
             return false;
         }
 
-        if (!$this->serverClient->get('has_voot_token', ['user_id' => $userInfo->id()])) {
+        if (false === $response = $this->client->get($this->provider, $userInfo->id(), 'groups', $this->vootUri)) {
+            // "false" is returned for a number of reasons:
+            // * no access_token yet for this user ID / scope
+            // * access_token expired (and no refresh_token available)
+            // * access_token was not accepted (revoked?)
+            // * refresh_token was rejected (revoked?)
+            //
+            // we need to re-request authorization at the OAuth server, redirect
+            // the browser to the authorization endpoint (with a 302)
             return new RedirectResponse(
                 sprintf(
                     '%s_voot/authorize?%s',
@@ -74,6 +120,40 @@ class VootTokenHook implements BeforeHookInterface
             );
         }
 
+        if (!$response->isOkay()) {
+            // VOOT responses should be HTTP 200 responses... there is
+            // something else wrong...
+            // XXX we should probably die here!
+            return false;
+        }
+
+        $vootGroups = self::extractMembership($response->json());
+        $this->session->set('_cached_voot_groups', $vootGroups);
+        $this->session->set('_cached_voot_groups_user_id', $userInfo->id());
+        $hookData['auth']->addEntitlements($vootGroups);
+
         return true;
+    }
+
+    /**
+     * @return array<string>
+     */
+    private static function extractMembership(array $responseData)
+    {
+        $memberOf = [];
+        foreach ($responseData as $groupEntry) {
+            if (!\is_array($groupEntry)) {
+                continue;
+            }
+            if (!array_key_exists('id', $groupEntry)) {
+                continue;
+            }
+            if (!\is_string($groupEntry['id'])) {
+                continue;
+            }
+            $memberOf[] = $groupEntry['id'];
+        }
+
+        return $memberOf;
     }
 }